Restructure.
