+++ src/main/java/com/autonomouslogic/everef/cli/refdata/esi/EsiTransformer.java
@@ -1,4 +1,4 @@
-package com.autonomouslogic.everef.refdata.esi;
+package com.autonomouslogic.everef.cli.refdata.esi;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
