+++ src/main/java/com/autonomouslogic/everef/cli/refdata/esi/EsiTypeTransformer.java
@@ -1,6 +1,6 @@
-package com.autonomouslogic.everef.refdata.esi;
+package com.autonomouslogic.everef.cli.refdata.esi;
 
-import com.autonomouslogic.everef.refdata.TransformUtil;
+import com.autonomouslogic.everef.cli.refdata.TransformUtil;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import javax.inject.Inject;
 import javax.inject.Singleton;
