+++ src/test/java/com/autonomouslogic/everef/cli/refdata/esi/EsiLoaderTest.java
@@ -1,4 +1,4 @@
-package com.autonomouslogic.everef.refdata.sde;
+package com.autonomouslogic.everef.cli.refdata.esi;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -14,11 +14,13 @@ import lombok.extern.log4j.Log4j2;
 import org.h2.mvstore.MVMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvFileSource;
 
 @Log4j2
-public class SdeLoaderTest {
+public class EsiLoaderTest {
 	@Inject
-	SdeLoader sdeLoader;
+	EsiLoader esiLoader;
 
 	@Inject
 	ObjectMapper objectMapper;
@@ -35,18 +37,24 @@ public class SdeLoaderTest {
 	@SneakyThrows
 	void before() {
 		DaggerTestComponent.builder().build().inject(this);
-		var mvstore = mvStoreUtil.createTempStore(SdeLoaderTest.class.getSimpleName());
+		var mvstore = mvStoreUtil.createTempStore(EsiLoaderTest.class.getSimpleName());
 		typeStore = mvStoreUtil.openJsonMap(mvstore, "types", Long.class);
-		sdeLoader.setTypeStore(typeStore);
+		esiLoader.setTypeStore(typeStore);
 	}
 
 	@Test
 	@SneakyThrows
-	void testLoadSde() {
-		sdeLoader.load(testDataUtil.createTestSde()).blockingAwait();
+	void testLoadEsi() {
+		esiLoader.load(testDataUtil.createTestEsiDump()).blockingAwait();
 		assertEquals(1, typeStore.size());
-		var expectedType = objectMapper.readTree(ResourceUtil.loadContextual(SdeLoaderTest.class, "/type-645.json"));
-		var actual = typeStore.get(645L);
-		testDataUtil.assertJsonStrictEquals(expectedType, actual);
+		var expectedType = objectMapper.readTree(ResourceUtil.loadContextual(EsiLoaderTest.class, "/type-645.json"));
+		testDataUtil.assertJsonStrictEquals(expectedType, typeStore.get(645L));
+	}
+
+	@ParameterizedTest
+	@CsvFileSource(resources = "/com/autonomouslogic/everef/cli/refdata/esi/EsiLoaderTest/filenames.csv")
+	void shouldParseFileNames(String filename, String fileType, String language) {
+		assertEquals(fileType, esiLoader.getFileType(filename));
+		assertEquals(language, esiLoader.getLanguage(filename));
 	}
 }
