+++ src/test/java/com/autonomouslogic/everef/test/TestComponent.java
@@ -7,6 +7,11 @@ import com.autonomouslogic.everef.cli.decorator.SlackDecoratorTest;
 import com.autonomouslogic.everef.cli.marketorders.ScrapeMarketOrdersTest;
 import com.autonomouslogic.everef.cli.publiccontracts.ScrapePublicContractsTest;
 import com.autonomouslogic.everef.cli.refdata.BuildRefDataTest;
+import com.autonomouslogic.everef.cli.refdata.FieldRenamerTest;
+import com.autonomouslogic.everef.cli.refdata.ObjectMergerTest;
+import com.autonomouslogic.everef.cli.refdata.TransformUtilTest;
+import com.autonomouslogic.everef.cli.refdata.esi.EsiLoaderTest;
+import com.autonomouslogic.everef.cli.refdata.sde.SdeLoaderTest;
 import com.autonomouslogic.everef.esi.EsiHelperTest;
 import com.autonomouslogic.everef.esi.EsiLimitExceededInterceptorTest;
 import com.autonomouslogic.everef.esi.EsiRateLimitInterceptorTest;
@@ -18,11 +23,6 @@ import com.autonomouslogic.everef.http.MockDataCrawlerModule;
 import com.autonomouslogic.everef.inject.AwsModule;
 import com.autonomouslogic.everef.inject.EsiModule;
 import com.autonomouslogic.everef.inject.JacksonModule;
-import com.autonomouslogic.everef.refdata.FieldRenamerTest;
-import com.autonomouslogic.everef.refdata.ObjectMergerTest;
-import com.autonomouslogic.everef.refdata.TransformUtilTest;
-import com.autonomouslogic.everef.refdata.esi.EsiLoaderTest;
-import com.autonomouslogic.everef.refdata.sde.SdeLoaderTest;
 import com.autonomouslogic.everef.url.UrlParserTest;
 import dagger.Component;
 import javax.inject.Singleton;
