+++ src/test/resources/com/autonomouslogic/everef/cli/refdata/ObjectMergerTest/sde-renames.csv
