+++ src/test/resources/com/autonomouslogic/everef/cli/refdata/esi/EsiLoaderTest/filenames.csv
