+++ src/test/resources/com/autonomouslogic/everef/cli/refdata/esi/EsiLoaderTest/type-645.json
