+++ src/test/resources/com/autonomouslogic/everef/cli/refdata/sde/SdeLoaderTest/type-645.json
